refactor: align fonts
