+++ ui/apps/dashboard/src/app/baseWrapper.tsx
@@ -1,16 +1,5 @@
-import Font from 'next/font/local';
-
+import { interTight, robotoMono } from '@/app/fonts';
 import cn from '@/utils/cn';
-
-const inter = Font({
-  src: '../fonts/InterTight-VariableFont.woff2',
-  variable: '--font-inter',
-});
-
-const robotoMono = Font({
-  src: '../fonts/RobotoMono-VariableFont.woff2',
-  variable: '--font-roboto-mono',
-});
 
 // This is separated from RootLayout so that we can use it in Storybook. For
 // whatever reason, <PageViewTracker /> makes Storybook error.
@@ -18,7 +7,7 @@ export function BaseWrapper({ children }: { children: React.ReactNode }) {
   return (
     <html
       lang="en"
-      className={cn('h-full bg-white accent-indigo-500', inter.variable, robotoMono.variable)}
+      className={cn('h-full bg-white accent-indigo-500', interTight.variable, robotoMono.variable)}
     >
       <body className="h-full overflow-hidden">{children}</body>
     </html>
