+++ ui/apps/dev-server-ui/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 
 import StoreProvider from '@/app/StoreProvider';
-import { robotoMono } from '@/app/fonts';
+import { interTight, robotoMono } from '@/app/fonts';
 import './globals.css';
 
 export const metadata: Metadata = {
@@ -14,7 +14,7 @@ type RootLayoutProps = {
 
 export default function RootLayout({ children }: RootLayoutProps) {
   return (
-    <html lang="en" className={`${robotoMono.variable}`}>
+    <html lang="en" className={`${interTight.variable} ${robotoMono.variable}`}>
       <body className="bg-slate-940">
         <div id="app" />
         <div id="modals" />
